docker-classifier/requirements.txt
flask>=0.12.3
onnxruntime
Pillow
NumPy
markdown